requirements.txt
click==7.1.2
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
